moves 404 checking for challenges to a helper function and fixes 403 check to be more accurate
